refactor: rename asset tests to rendering_method tests
