+++ loco-new/tests/templates/mod.rs
@@ -9,7 +9,7 @@ use loco::{
 };
 use rand::{rngs::StdRng, SeedableRng};
 
-mod asset;
+
 mod auth;
 mod background;
 mod db;
@@ -17,6 +17,7 @@ mod features;
 mod initializers;
 mod mailer;
 mod module_name;
+mod rendering_method;
 
 pub struct TestGenerator {
     tree: tree_fs::Tree,
