+++ loco-new/tests/templates/rendering_method.rs
